Moved unleash config into separate file
