+++ packages/commonwealth/client/scripts/App.tsx
@@ -1,5 +1,5 @@
 import { OpenFeatureProvider } from '@openfeature/react-sdk';
-import { InMemoryProvider, OpenFeature } from '@openfeature/web-sdk';
+import { OpenFeature } from '@openfeature/web-sdk';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import useInitApp from 'hooks/useInitApp';
@@ -8,9 +8,7 @@ import React, { StrictMode } from 'react';
 import { RouterProvider } from 'react-router-dom';
 import { ToastContainer } from 'react-toastify';
 import { queryClient } from 'state/api/config';
-import { UnleashClient } from 'unleash-proxy-client';
-import { UnleashProvider } from '../../shared/UnleashProvider';
-import { featureFlags } from './helpers/feature-flags';
+import { openFeatureProvider } from './helpers/feature-flags';
 import { CWIcon } from './views/components/component_kit/cw_icons/cw_icon';
 
 const Splash = () => {
@@ -22,18 +20,7 @@ const Splash = () => {
   );
 };
 
-const unleashConfig = {
-  url: process.env.UNLEASH_FRONTEND_SERVER_URL,
-  clientKey: process.env.UNLEASH_FRONTEND_API_TOKEN,
-  refreshInterval: 120,
-  appName: 'commonwealth-web',
-};
-
-OpenFeature.setProvider(
-  process.env.UNLEASH_FRONTEND_API_TOKEN
-    ? new UnleashProvider(new UnleashClient(unleashConfig))
-    : new InMemoryProvider(featureFlags),
-);
+OpenFeature.setProvider(openFeatureProvider);
 
 const App = () => {
   const { customDomain, isLoading } = useInitApp();
